# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FoodInsight API",
    "description": "\n## Smart Snack Inventory Monitoring API\n\nLocal-first API for managing snack inventory on FoodInsight edge devices.\n\n### Architecture\n\n- **Local SQLite Database**: All data stored locally, no cloud dependencies\n- **Role-Based Access**: Public inventory endpoints, authenticated admin endpoints\n- **Real-Time Detection**: Receives inventory updates from edge detection pipeline\n\n### Authentication\n\nAdmin endpoints require HTTP Basic authentication:\n- `viewer`: Read-only access to status and configuration\n- `operator`: Can start/stop detection pipeline\n- `admin`: Full access including user management\n\n### Default Credentials\n\n| Username | Password | Role |\n|----------|----------|------|\n| admin | admin | admin |\n\n**Important**: Change default credentials after first login.\n        ",
    "version": "1.0.0",
    "x-logo": {
      "url": "https://example.com/logo.png"
    },
    "contact": {
      "name": "FoodInsight Support",
      "email": "support@foodinsight.local"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/inventory/events": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Inventory Events",
        "description": "Get recent inventory events.\n\nPublic endpoint - no authentication required.",
        "operationId": "get_inventory_events_inventory_events_get",
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "item_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item name",
              "title": "Item Name"
            },
            "description": "Filter by item name"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Device Status",
        "description": "Get current device status.",
        "operationId": "get_device_status_admin_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/admin/config": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Config",
        "description": "Get all device configuration.",
        "operationId": "get_config_admin_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Config",
        "description": "Update a configuration value.",
        "operationId": "update_config_admin_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/admin/detection/start": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Start Detection",
        "description": "Start the detection pipeline.",
        "operationId": "start_detection_admin_detection_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/admin/detection/stop": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Stop Detection",
        "description": "Stop the detection pipeline.",
        "operationId": "stop_detection_admin_detection_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/admin/events": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Events",
        "description": "Get detection events with filters.",
        "operationId": "get_events_admin_events_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "item_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigUpdate": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "ConfigUpdate",
        "description": "Configuration update request."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}